+++ src/AspNetCoreRateLimit/Core/IRateLimitProcessor.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace AspNetCoreRateLimit
+{
+    public interface IRateLimitProcessor
+    {
+        IEnumerable<RateLimitRule> GetMatchingRules(ClientRequestIdentity identity);
+
+        RateLimitHeaders GetRateLimitHeaders(ClientRequestIdentity requestIdentity, RateLimitRule rule);
+
+        bool IsWhitelisted(ClientRequestIdentity requestIdentity);
+
+        RateLimitCounter ProcessRequest(ClientRequestIdentity requestIdentity, RateLimitRule rule);
+    }
+}
