+++ src/AspNetCoreRateLimit/Models/ClientRateLimitOptions.cs
@@ -1,6 +1,6 @@
 ﻿namespace AspNetCoreRateLimit
 {
-    public class ClientRateLimitOptions : RateLimitCoreOptions
+    public class ClientRateLimitOptions : RateLimitOptions
     {
         /// <summary>
         /// Gets or sets the HTTP header that holds the client identifier, by default is X-ClientId
