+++ src/AspNetCoreRateLimit/Models/RateLimitPolicy.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace AspNetCoreRateLimit
+{
+    public class RateLimitPolicy
+    {
+        public List<RateLimitRule> Rules { get; set; }
+    }
+}
