+++ src/AspNetCoreRateLimit/Store/DistributedCacheClientPolicyStore.cs
@@ -8,18 +8,22 @@ namespace AspNetCoreRateLimit
     {
         private readonly IDistributedCache _memoryCache;
 
-        public DistributedCacheClientPolicyStore(IDistributedCache memoryCache, 
+        public DistributedCacheClientPolicyStore(
+            IDistributedCache memoryCache, 
             IOptions<ClientRateLimitOptions> options = null, 
             IOptions<ClientRateLimitPolicies> policies = null)
         {
             _memoryCache = memoryCache;
 
-            //save client rules defined in appsettings in distributed cache on startup
-            if (options != null && options.Value != null && policies != null && policies.Value != null && policies.Value.ClientRules != null)
+            var clientOptions = options?.Value;
+            var clientPolicyRules = policies?.Value?.ClientRules;
+
+            //save client rules defined in appsettings in cache on startup
+            if (clientOptions != null && clientPolicyRules != null)
             {
-                foreach (var rule in policies.Value.ClientRules)
+                foreach (var rule in clientPolicyRules)
                 {
-                    Set($"{options.Value.ClientPolicyPrefix}_{rule.ClientId}", new ClientRateLimitPolicy { ClientId = rule.ClientId, Rules = rule.Rules });
+                    Set($"{clientOptions.ClientPolicyPrefix}_{rule.ClientId}", new ClientRateLimitPolicy { ClientId = rule.ClientId, Rules = rule.Rules });
                 }
             }
         }
@@ -32,16 +36,19 @@ namespace AspNetCoreRateLimit
         public bool Exists(string id)
         {
             var stored = _memoryCache.GetString(id);
+
             return !string.IsNullOrEmpty(stored);
         }
 
         public ClientRateLimitPolicy Get(string id)
         {
             var stored = _memoryCache.GetString(id);
+
             if (!string.IsNullOrEmpty(stored))
             {
                 return JsonConvert.DeserializeObject<ClientRateLimitPolicy>(stored);
             }
+
             return null;
         }
