+++ src/AspNetCoreRateLimit/Store/DistributedCacheIpPolicyStore.cs
@@ -8,16 +8,21 @@ namespace AspNetCoreRateLimit
     {
         private readonly IDistributedCache _memoryCache;
 
-        public DistributedCacheIpPolicyStore(IDistributedCache memoryCache, 
+        public DistributedCacheIpPolicyStore(
+            IDistributedCache memoryCache, 
             IOptions<IpRateLimitOptions> options = null, 
             IOptions<IpRateLimitPolicies> policies = null)
         {
             _memoryCache = memoryCache;
 
-            //save ip rules defined in appsettings in distributed cache on startup
-            if (options != null && options.Value != null && policies != null && policies.Value != null && policies.Value.IpRules != null)
+            var ipOptions = options?.Value;
+            var ipPolicyRules = policies?.Value;
+
+            //save IP rules defined in appsettings in cache on startup
+            if (ipOptions != null && ipPolicyRules != null)
             {
-                Set($"{options.Value.IpPolicyPrefix}", policies.Value);
+                Set($"{ipOptions.IpPolicyPrefix}", ipPolicyRules);
+
             }
         }
