+++ src/AspNetCoreRateLimit/Store/IClientPolicyStore.cs
@@ -1,10 +1,6 @@
 ﻿namespace AspNetCoreRateLimit
 {
-    public interface IClientPolicyStore
+    public interface IClientPolicyStore : IPolicyStore<ClientRateLimitPolicy>
     {
-        bool Exists(string id);
-        ClientRateLimitPolicy Get(string id);
-        void Remove(string id);
-        void Set(string id, ClientRateLimitPolicy policy);
     }
 }
