+++ src/AspNetCoreRateLimit/Store/IIpPolicyStore.cs
@@ -1,10 +1,6 @@
 ﻿namespace AspNetCoreRateLimit
 {
-    public interface IIpPolicyStore
+    public interface IIpPolicyStore : IPolicyStore<IpRateLimitPolicies>
     {
-        bool Exists(string id);
-        IpRateLimitPolicies Get(string id);
-        void Remove(string id);
-        void Set(string id, IpRateLimitPolicies policy);
     }
 }
