+++ src/AspNetCoreRateLimit/Store/IPolicyStore.cs
@@ -1,0 +1,10 @@
+﻿namespace AspNetCoreRateLimit
+{
+    public interface IPolicyStore<TPolicy>
+    {
+        bool Exists(string id);
+        TPolicy Get(string id);
+        void Remove(string id);
+        void Set(string id, TPolicy policy);
+    }
+}
